Refactor BLE callbacks
